requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Document Processing
pymupdf
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=10.0.0

# NLP & ML
spacy>=3.7.0
transformers>=4.37.0
torch>=2.2.0
sentence-transformers>=2.2.2

# Data Generation & Analysis
faker>=22.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# File Watching & Notifications (NEW)
watchdog>=3.0.0
plyer>=2.1.0

# Encryption (NEW)
cryptography>=42.0.0

# Document Processing for File Watcher (NEW)
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
